requirements.txt
# Telegram Bot
aiogram>=3.13.0

# Scheduler
apscheduler>=3.10.4

# Environment & Config
python-dotenv>=1.0.0

# Database (SQLite via aiosqlite for async)
aiosqlite>=0.19.0

# Optional: for production logging
python-json-logger>=2.0.7
